#61 id를 이용해 정보를 불러오는 부분 추가 조건

+++ src/main/java/api/damdev/moneybook/controller/CycleController.java
@@ -52,6 +52,10 @@ public class CycleController {
 
     @DeleteMapping("{cycleId}")
     public ResponseEntity removeCycle(@PathVariable("cycleId") String id) {
+        if(Strings.isNullOrEmpty(id)) {
+            return ResponseEntity.badRequest().build();
+        }
+
         Cycle cycle = cycleService.deleteCycle(id);
 
         if(Strings.isNullOrEmpty(cycle.getId())) {
